TC02-027 Emit attribute RAC:assume
The function calls and function guards in the postcondition of
function declarations cannot be proven because they lack a
definition, resulting in incomplete RAC.

This change emits the label [@RAC:assume] for these conjuncts of the
postcondition, which are then added to the assumptions when checking
program annotation during RAC in gnatwhy3.

* src/common/vc_kinds.ads
Add definition of RAC_Assume_Label
* src/why/gnat2why-subprograms.adb
Use RAC_Assume_Label in postconditions of function declarations.

Change-Id: I2b3fd01b8fff38b469dccec25532b12e2850deff